==> BAYSM_2024/04_BRMS_Model_Estimates.Rmd <==
---
title: "BRMS Model estimates"
author: "Oluwayomi Akinfenwa"
date: "08/03/2024"
output:
  html_document:
    toc: yes
    toc_depth: '4'
    df_print: paged
  bookdown::html_document2:
    toc: no
    toc_float: yes
    toc_depth: 4
    number_sections: yes
  pdf_document:
    toc: yes
    toc_depth: '4'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
options(knitr.duplicate.label = 'allow')
```

### Loading necessary packages

```{r, libraries, include = FALSE}
library(stringr)
library(tidybayes)
library(ggplot2)
library(tidyr)
library(dplyr)
library(brms)
library(bayesplot)
library(patchwork)
library(ggragged)
library(geofacet)
library(grid)
```

Loading the PISA data for the European countries - Europe_Pisamaths.Rdata.

```{r loading the data}
load(here::here("PISA_Data", "Europe_Pisamaths.Rdata"))

# We have decided to remove year 2022 from my data and make prediction for 2022.
# Hence i will filter year = 2022.
PISA_Europe_Data <- Pisa_Europe_Data |>
  filter(year != "2022")
```

### Creating the grid for the prediction

```{r}
country_names <- levels(factor(PISA_Europe_Data$Country))
region_names <- levels(factor(PISA_Europe_Data$Region))
income_names <- levels(factor(PISA_Europe_Data$Income))
incomeregion_names <- levels(factor(PISA_Europe_Data$Income_Region))

## Obtaining the index for the data
mu_index <- 1:nrow(PISA_Europe_Data)

## Obtaining the region indexes
country_region <- PISA_Europe_Data |> group_by(Country) |>
  summarise(Region=first(Region)) |>
  mutate(Country_num = as.numeric(as.factor(Country))) |>
  mutate(region_num = as.numeric(as.factor(Region)))

REGION <- country_region |> select(Country, Region)

## Obtaining the income indexes
country_income <- PISA_Europe_Data |> group_by(Country) |>
  summarise(Income=first(Income)) |>
  mutate(Country_num = as.numeric(as.factor(Country))) |>
  mutate(income_num = as.numeric(as.factor(Income)))
         
INCOME <- country_income |> select(Country, Income)

## Obtaining the income_region indexes
country_incomeregion <- PISA_Europe_Data |> group_by(Country) |>
  summarise(Income_Region=first(Income_Region)) |>
  mutate(Country_num = as.numeric(as.factor(Country))) |>
  mutate(incomeregion_num = as.numeric(as.factor(Income_Region)))

INCOME_REGION <- REGION |>
  left_join(INCOME, join_by(Country))

Country_region <- country_region |> left_join(INCOME, join_by(Country))
Country_income <- country_income |> left_join(REGION, join_by(Country))
Country_incomeregion <- country_incomeregion |> 
  left_join(INCOME_REGION, join_by(Country))
```


## Loading the hierarchical models from the saved R.Data file

```{r}
#Independent-Country specific model
load(here::here("Manuscript_Models", "CountryInd_BRMSModel.Rdata"))

#Country specific model
load(here::here("Manuscript_Models", "Country_BRMSModel.Rdata"))

#Region hierarchical model
load(here::here("Manuscript_Models", "Region_BRMSModel.Rdata"))

#Income hierarchical model
load(here::here("Manuscript_Models", "Income_BRMSModel.Rdata"))

#Income-Region hierarchical model
load(here::here("Manuscript_Models", "IncomeRegion_BRMSModel.Rdata"))
```

### Pulling the estimates from all the fitted models

#### Estimates from the independent-country-specific model

```{r}
#my observations
# the global estimates is the estimate for Albania. Albania is used as the base country.
# also Bosnia was messed up and unitedkingdom merged together.
#hence, the str_replace_all might not work to fix this and I have to fix this manually.

CountryInd_draws <-posterior::as_draws_rvars(CountryInd_BRMSModel, 
                variable = "^b_Country|^b_Intercept",regex = TRUE) |>
  purrr::map_dfr(as_tibble, .id = "name")|>
  mutate(rvar=value+value[1])

#Albania is used as the base country, hence the intercept estimate is the rvar for Albania.
CountryInd_draws$rvar[1] <- CountryInd_draws$value[1]
  
  CountryInd_I <- CountryInd_draws |>
  mutate(Country = stringr::str_replace_all(name, stringr::fixed("b_Country"), "")) |>
  mutate(Country = stringr::str_replace_all(Country, stringr::fixed("b_Intercept"), "Albania")) |>
  mutate(Country = stringr::str_replace_all(Country, stringr::fixed("UnitedKingdom"), "United Kingdom")) |>
  right_join(Country_region, join_by(Country== Country))|>
    mutate(term = "year_orig") |>
  select(Country,Region, Income, term, rvar)

#Slope estimates
CountrySlop_draws <- posterior::as_draws_rvars(CountryInd_BRMSModel, 
                    variable = "^b_year_orig",regex = TRUE) |>
  purrr::map_dfr(as_tibble, .id = "name")|>
  mutate(rvar=value+value[1])

#Albania is used as the base country, hence the intercept estimate is the rvar for Albania.
CountrySlop_draws$rvar[1] <- CountrySlop_draws$value[1]

CountryInd_S <- CountrySlop_draws |>
  mutate(Country = stringr::str_replace_all(name, stringr::fixed("b_year_orig:Country"), "")) |>
  mutate(Country = stringr::str_replace_all(Country, stringr::fixed("b_year_orig"), "Albania")) |>
  mutate(Country = stringr::str_replace_all(Country, stringr::fixed("UnitedKingdom"), "United Kingdom")) |>
  right_join(Country_region, join_by(Country== Country))|>
  mutate(term = "year_orig") |>
  select(Country,Region, Income, term, rvar)

## Tidying the results
CountryInd_Int <- CountryInd_I |> mutate(Intercept= median(rvar)) |>
  select(Country, Region, Intercept)
CountryInd_Slop <- CountryInd_S |> mutate(year_orig= median(rvar)) |>
  select(Country, Region, year_orig)

CountryInd_IS_Est <- CountryInd_Int |>
  left_join(CountryInd_Slop, join_by(Country, Region))

CountryInd_Int_Est <- CountryInd_I |>
  mutate(Model = "Independent \nModel") |>
  select(Country, Model, rvar)

CountryInd_Slop_Est <- CountryInd_S|>
  mutate(Model = "Independent \nModel") |>
  select(Country, Model, rvar)
```

#### Estimates from the country-specific model

```{r}
## Pulling the intercept results. The global intercept (b_Intercept) and the country offsets.
Country_I <- spread_rvars(Country_BRMSModel, r_Country[Country,term], b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(Country = stringr::str_replace_all(Country, fixed("."), " ")) |>
  mutate(r_Country = r_Country+ b_Intercept) |>
  rename(rvar=r_Country) |>
  left_join(Country_region, join_by(Country== Country))

Country_Int <- Country_I|>
  select(Country,Region, term, rvar)

## Pulling the slope results. The global slope (b_year_orig) and the country offsets.
Country_S <- spread_rvars(Country_BRMSModel, r_Country[Country,term], b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(Country = stringr::str_replace_all(Country, fixed("."), " ")) |>
  mutate(r_Country = r_Country+ b_year_orig) |>
   rename(rvar=r_Country) |>
  left_join(Country_region, join_by(Country== Country))

Country_Slop <- Country_S|>
  select(Country,Region, term, rvar)

## Tidying the estimates
Country_IS <- rbind(Country_Int, Country_Slop)

Country_IS_Est <- Country_IS |> mutate(coef=median(rvar))|> 
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

## Intercept Estimates
Country_Int_Est <- Country_I |>
  mutate(Model = "Country \nModel") |>
  select(c(Country, Model, rvar))

## Slope Estimates
Country_Slop_Est <- Country_S |>
  mutate(Model = "Country \nModel") |>
  select(c(Country,Model, rvar))

#The hierarchical estimate for the country-specific model is the b_intercept (global estimate)
Country_Int_est <- Country_I |>
  mutate(Model = "Country \nModel") |>
  select(c(Country, Model,rvar = b_Intercept))

Country_Slop_est <- Country_S |>
  mutate(Model = "Country \nModel") |>
  select(c(Country, Model,rvar= b_year_orig))
```

#### Estimates from the Region hierarchical model

```{r}
Region_int <- spread_rvars(Region_BRMSModel, r_Region[group1,term], b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " ")) |>
  mutate(r_Region = r_Region+ b_Intercept) |>
  rename(rvar=r_Region, Region = group1) |>
  select(Region, term, rvar)

Region_slop <- spread_rvars(Region_BRMSModel, r_Region[group1,term], b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " ")) |>
  mutate(r_Region = r_Region+  b_year_orig) |>
  rename(rvar=r_Region, Region = group1)|>
  select(Region, term, rvar)

Region_is <- rbind(Region_int,Region_slop) 

#Extending the estimates across all the countries

Region_Int <- Region_int |> 
  right_join(Country_region, by = join_by(Region == Region)) |>
  select(Country,Region, Income, term, rvar)

Region_Slop <- Region_slop |> 
  right_join(Country_region, by = join_by(Region == Region)) |>
  select(Country,Region, Income, term, rvar)

Region_IS <- rbind(Region_Int, Region_Slop)
  
## The region country estimates

Rcountry_Int <- spread_rvars(Region_BRMSModel, r_Region[group1,term],r_Country[group,term],b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " "), 
         group = stringr::str_replace_all(group, fixed("."), " ")) |>
  right_join(Country_region, by = join_by(group==Country,group1==Region)) |> #this join everything from the country_region with its intercession from the model result
  mutate(r_Country =  r_Country+ r_Region+ b_Intercept) |>
  rename(rvar=r_Country,Region=group1, Country =group) |>
  select(Country,Region, Income, term, rvar)

Rcountry_Slop <- spread_rvars(Region_BRMSModel, r_Region[group1,term],r_Country[group,term],b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " "), 
         group = stringr::str_replace_all(group, fixed("."), " ")) |>
  right_join(Country_region, by = join_by(group==Country,group1==Region)) |>
  mutate(r_Country =  r_Country+ r_Region+ b_year_orig) |>
  rename(rvar=r_Country,Region=group1, Country =group) |>
  select(Country,Region, Income, term, rvar)

Rcountry_IS <- rbind(Rcountry_Int, Rcountry_Slop) 
```

### Tidying the estimates from the Region hierarchical model

```{r}
#median estimates
Region_IS_Est <- Region_IS |> mutate(coef=median(rvar))|>
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

Rcountry_IS_Est <- Rcountry_IS |> mutate(coef=median(rvar))|> 
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

## Credible interval
## Intercept Estimates
#Hierarchy estimates
Region_Int_est <- Region_Int |>
  mutate(Model = "Region \nModel")|>
  select(Country, Model,rvar)
#Country estimates
Region_Int_Est <- Rcountry_Int |> 
  mutate(Model = "Region \nModel")|>
  select(c(Country,Model,rvar))
## Slope Estimates
#Hierarchy estimates
Region_Slop_est <- Region_Slop |>
  mutate(Model = "Region \nModel")|>
  select(c(Country,Model,rvar))
#Country estimates
Region_Slop_Est <- Rcountry_Slop |> 
  mutate(Model = "Region \nModel")|>
  select(Country, Model,rvar)
```

#### Estimates from the income hierarchical model

```{r}
Income_int <- spread_rvars(Income_BRMSModel, r_Income[group1,term],b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " ")) |>
  mutate(r_Income = r_Income+ b_Intercept) |>
  rename(rvar=r_Income, Income= group1) |>
  select(Income, term, rvar)

Income_slop <- spread_rvars(Income_BRMSModel, r_Income[group1,term], b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " ")) |>
  mutate(r_Income = r_Income+  b_year_orig) |>
  rename(rvar=r_Income, Income= group1)|>
  select(Income, term, rvar)
Income_is <- rbind(Income_int, Income_slop)

#Extending the income hierarchical estimates across the countries.
Income_Int <- Income_int |>
  right_join(Country_income, by = join_by(Income ==Income))|>
  select(Country, Region, Income, term, rvar)

Income_Slop <- Income_slop |>
  right_join(Country_income, by = join_by(Income ==Income))|>
  select(Country, Region, Income, term, rvar)

Income_IS <- rbind(Income_Int, Income_Slop)
```

pulling the intercepts and slopes for each country based on the influence of the Income hierarchy

```{r}
Icountry_Int <- spread_rvars(Income_BRMSModel, r_Income[group1,term],r_Country[group,term],b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " "), 
         group = stringr::str_replace_all(group, fixed("."), " ")) |>
  right_join(Country_income, by = join_by(group==Country,group1==Income)) |> #this join everything from the country_Income with its intercession from the model result
  mutate(r_Country =  r_Country+ r_Income+ b_Intercept) |>
  rename(rvar=r_Country, Country = group, Income = group1) |>
  select(Country, Income,Region, term, rvar)


Icountry_Slop <- spread_rvars(Income_BRMSModel, r_Income[group1,term],r_Country[group,term],b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " "), 
         group = stringr::str_replace_all(group, fixed("."), " ")) |>
  right_join(Country_income, by = join_by(group==Country,group1==Income)) |>
  mutate(r_Country =  r_Country+ r_Income+ b_year_orig) |>
  rename(rvar=r_Country, Country = group, Income = group1) |>
  select(Country, Income, Region, term, rvar)

Icountry_IS <- rbind(Icountry_Int, Icountry_Slop) 
```

#### Tidying the estimates from the income model

```{r, fig.width=10, fig.height=8}
#Hierarchical estimates
Income_IS_EST <- Income_IS |> mutate(coef=median(rvar))|> 
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

#Country-estimates based on the hierarchical pulling.
Icountry_IS_Est <- Icountry_IS |> mutate(coef=median(rvar))|> 
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

## Credible interval
## Intercept Estimates
#Hierarchy estimates
Income_Int_est <- Income_Int |> 
  mutate(Model = "Income \nModel")|>
  select(c(Country,Model,rvar))
#Country estimates
Income_Int_Est <- Icountry_Int |> 
  mutate(Model = "Income \nModel")|>
  select(Country, Model,rvar)
## Slope Estimates
#Hierarchy estimates
Income_Slop_est <- Income_Slop |> 
  mutate(Model = "Income \nModel")|>
  select(c(Country,Model,rvar))
#Country estimates
Income_Slop_Est <- Icountry_Slop |> 
  mutate(Model = "Income \nModel")|>
  select(Country, Model,rvar)
```

#### Estimates from the IncomeRegion hierarchical model

```{r}
IncomeRegion_int <- spread_rvars(IncomeRegion_BRMSModel, r_Income_Region[group1,term],b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " ")) |>
  mutate(r_Income_Region = r_Income_Region+ b_Intercept) |>
  rename(rvar=r_Income_Region, `Income_Region` = group1) |>
  select(Income_Region, term, rvar)

IncomeRegion_slop <- spread_rvars(IncomeRegion_BRMSModel, r_Income_Region[group1,term], b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " ")) |>
  mutate(r_Income_Region = r_Income_Region+  b_year_orig) |>
  rename(rvar=r_Income_Region, `Income_Region` = group1) |>
  select(Income_Region, term, rvar)

IncomeRegion_is <- rbind(IncomeRegion_int, IncomeRegion_slop)

#Extending the income-region hierarchical estimates across the countries.
IncomeRegion_Int <- IncomeRegion_int |>
  right_join(Country_incomeregion, by = join_by(Income_Region ==Income_Region))|>
  select(Country, Income_Region, term, rvar)

IncomeRegion_Slop <- IncomeRegion_slop |>
  right_join(Country_incomeregion, by = join_by(Income_Region ==Income_Region))|>
  select(Country, Income_Region, term, rvar)


IncomeRegion_IS <- rbind(IncomeRegion_Int, IncomeRegion_Slop)
```

pulling the intercepts and slopes for each country based on the influence of the Income-Region hierarchy.

```{r}
IRcountry_Int <- spread_rvars(IncomeRegion_BRMSModel, r_Income_Region[group1,term],r_Country[group,term],b_Intercept) |>
  filter(term == "Intercept") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " "), 
         group = stringr::str_replace_all(group, fixed("."), " ")) |>
  right_join(Country_incomeregion, by = join_by(group==Country,group1==Income_Region)) |> #this join everything from the country_Income with its intercession from the model result
  mutate(r_Country =  r_Country+ r_Income_Region+ b_Intercept) |>
  rename(rvar=r_Country, Country = group, Income_Region = group1) |>
  select(Country, Income_Region,Region, term, rvar)


IRcountry_Slop <- spread_rvars(IncomeRegion_BRMSModel, r_Income_Region[group1,term],r_Country[group,term],b_year_orig) |>
  filter(term == "year_orig") |>
  mutate(group1 = stringr::str_replace_all(group1, fixed("."), " "), 
         group = stringr::str_replace_all(group, fixed("."), " ")) |>
  right_join(Country_incomeregion, by = join_by(group==Country,group1==Income_Region)) |>
  mutate(r_Country =  r_Country+ r_Income_Region+ b_year_orig) |>
  rename(rvar=r_Country, Country = group, Income_Region = group1) |>
  select(Country, Income_Region, Region, term, rvar)

IRcountry_IS <- rbind(IRcountry_Int, IRcountry_Slop) 
```

#### Tidying the estimates from the model

```{r, fig.width=10, fig.height=8}
IR_IS_EST <- IncomeRegion_IS |> mutate(coef=median(rvar))|> 
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

IRcountry_IS_Est <- IRcountry_IS |> mutate(coef=median(rvar))|> 
  select(-rvar)|>
  pivot_wider(values_from=coef, names_from=term)

## Credible interval
## Intercept Estimates
#Hierarchy estimates
IncomeRegion_Int_est <- IncomeRegion_Int |> 
  mutate(Model = "Income-Region \nModel")|>
  select(Country,Model,rvar)
#Country estimates
Income_Region_Int_Est <- IRcountry_Int |> 
  mutate(Model = "Income-Region \nModel")|>
  select(Country, Model,rvar)
## Slope Estimates
#Hierarchy estimates
IncomeRegion_Slop_est <- IncomeRegion_Slop |> 
  mutate(Model = "Income-Region \nModel")|>
  select(Country, Model,rvar)
#Country estimates
Income_Region_Slop_Est <- IRcountry_Slop |> 
  mutate(Model = "Income-Region \nModel")|>
  select(Country, Model,rvar)
```


### Plotting the regression model fit on the data

```{r}
if (!dir.exists(here::here("Saved_Plots/Regression_fits")))
 dir.create(here::here("Saved_Plots/Regression_fits"))
```


#### Independent country model fit on the data

```{r}
pdf("Saved_Plots/Regression_fits/BRMS_independent_fit.pdf",
    width = 11,
    height = 8)
ggplot(PISA_Europe_Data, aes(x=year_orig, y=math))+ 
  geom_point()+
  facet_wrap(~Country)+
  geom_abline(data=CountryInd_IS_Est, aes(slope=year_orig, intercept= Intercept), color="black")+
  scale_x_continuous(breaks = seq(-15,0,3),labels = function(x) x+2018) +
  xlab("year")+ ggtitle("Independent country model fit of average maths scores over time")+
  theme_bw()+
  theme(panel.spacing=unit(1,"lines"),
        legend.position = "bottom",
        plot.subtitle = element_text(family = "Consolas"),
        axis.text.x = element_text(angle = 45, hjust = 0.5, vjust = 0.5))

dev.off()
```

#### Country model fit on the data

```{r}
pdf("Saved_Plots/Regression_fits/BRMS_country_specific_fit.pdf",
    width = 11,
    height = 8)

ggplot(PISA_Europe_Data, aes(x=year_orig, y=math))+ 
  geom_point()+
  facet_wrap(~Country)+
  geom_abline(data=Country_IS_Est, aes(slope=year_orig, intercept= Intercept), color="magenta3")+
  scale_x_continuous(breaks = seq(-15,0,3),labels = function(x) x+2018) +
  xlab("year")+ ggtitle("Country model fit of average maths scores over time")+
  theme_bw()+
  theme(panel.spacing=unit(1,"lines"),
        legend.position = "bottom",
        plot.subtitle = element_text(family = "Consolas"),
        axis.text.x = element_text(angle = 45, hjust = 0.5, vjust = 0.5))
dev.off()
```

#### Region model fit on the data

```{r}
pdf("Saved_Plots/Regression_fits/BRMS_Region_hierarchical_fit.pdf",
    width = 11,
    height = 8)
ggplot(PISA_Europe_Data, aes(x=year_orig, y=math))+ 
  geom_point()+
  ggh4x::facet_nested_wrap(vars(Region, Country), nrow = 5)+
  geom_abline(data= Rcountry_IS_Est, aes(slope=year_orig, intercept= Intercept), color="blue3")+
  scale_x_continuous(breaks = seq(-15,0,3),labels = function(x) x+2018) +
  xlab("year")+ ggtitle("Region model fit of average maths scores over time")+
  theme_bw()+
  theme(panel.spacing=unit(1,"lines"),
        legend.position = "bottom",
        plot.subtitle = element_text(family = "Consolas"),
        axis.text.x = element_text(angle = 45, hjust = 0.5, vjust = 0.5))
dev.off()
```

#### Income model fit on the data

```{r}
pdf("Saved_Plots/Regression_fits/BRMS_Income_hierarchical_fit.pdf",
    width = 11,
    height = 8)
ggplot(PISA_Europe_Data, aes(x=year_orig, y=math))+ 
  geom_point()+
  ggh4x::facet_nested_wrap(vars(Income, Country), nrow = 5)+
  geom_abline(data= Icountry_IS_Est, aes(slope=year_orig, intercept= Intercept), color="red3")+
  scale_x_continuous(breaks = seq(-15,0,3),labels = function(x) x+2018) +
  xlab("year")+ ggtitle("Income model fit of average maths scores over time")+
  theme_bw()+
  theme(panel.spacing=unit(1,"lines"),
        legend.position = "bottom",
        plot.subtitle = element_text(family = "Consolas"),
        axis.text.x = element_text(angle = 45, hjust = 0.5, vjust = 0.5))
dev.off()
```

#### IncomeRegion model fit on the data

```{r}
pdf("Saved_Plots/Regression_fits/BRMS_Income-Region_hierarchical_fit.pdf",
    width = 11,
    height = 8)
ggplot(PISA_Europe_Data, aes(x=year_orig, y=math))+ 
  geom_point()+
  ggh4x::facet_nested_wrap(vars(Income_Region, Country), nrow = 5)+
  geom_abline(data= IRcountry_IS_Est, aes(slope=year_orig, intercept= Intercept), color="red3")+
  scale_x_continuous(breaks = seq(-15,0,3),labels = function(x) x+2018) +
  xlab("year")+ ggtitle("Income model fit of average maths scores over time")+
  theme_bw()+
  theme(panel.spacing=unit(1,"lines"),
        legend.position = "bottom",
        plot.subtitle = element_text(family = "Consolas"),
        axis.text.x = element_text(angle = 45, hjust = 0.5, vjust = 0.5))
dev.off()
```
